fix: satisfy backend analytics test contracts
